+++ src/components/composite/entities/User/UserAvatar.tsx
